Define PrivateIface of transports in separate header
Summary:
No functional changes. It just moves the definition of a nested class into a separate header.

The benefit is that while the "enclosing" class is a public class (exposed to users such as PyTorch), the nested class is a private interface. By separating them we clarify this, and we avoid exposing internal details to external users. This latter aspect is meant both "in principle" and on a technical level: the private interface sometimes contains references to some implementation details (e.g., libuv or libibverbs structs) which therefore force us to include these headers in the public headers and thus propagate them to our end users.

This is also a first step towards getting rid of the separation between Impls and PrivateIfaces (this is why the new file is called `context_impl.h`), although this will happen later.

I don't remember why we hadn't done something of this sort from the very beginning. Maybe we didn't want to add a new file (even though it's not a big deal). Maybe we just didn't think of it...

Reviewed By: patricklabatut

Differential Revision: D24361244

fbshipit-source-id: 9fe4b421d093761086ccbd9dd01e2e6734f80358