Move the transport/channel registries into the benchmarks
Summary:
The transport and channel registries were introduced in order to support "distributed registration" for benchmarks and, in the end, kept being used only there. (the PyTorch RPC agent has its own separate registry, in its own codebase). Moreover, the registry has recently started to show some limits: we had never figured out how to support backends with parameters (such as the MPT channel) and we didn't support CUDA channels.

The benchmarks still need it, but let's "cordon" it off to that part of the codebase, to avoid it getting in the way of the main library code.

Reviewed By: beauby

Differential Revision: D24362589

fbshipit-source-id: 6632bb24e0db838aa6eb5304db7ad7d856a4f8cf