Use make_tuple instead of implicitly constructing tuples
Summary: In the recent IBV code I added several instances of a tuple being implicitly constructed in the return statement through an initializer list. This doesn't work in GCC 5.4 (apparently due to a defect in the standard itself that was later addressed, thanks beauby for finding that out) which PyTorch uses. Hence we move all of them to std::make_tuple.

Reviewed By: mrshenli

Differential Revision: D24539276

fbshipit-source-id: 34b299f9f324a655b6d4afb1fcc0f625cb57696d