Deduplicate logic replacing closed event loop with on-demand loop
Summary: All of our transports have an event loop thread which, over time, ended up acquiring two duties: respond to the events proper to the transport (the epoll in UV, the reactor in SHM, the completion queue poll in IBV) and accept and run deferred functions. We needed some logic in order to accomplish both these jobs, especially to deal with function deferral when the event loop was shutting down. This logic ended up being duplicated for all transports. Encapsulating it in its own class makes it much easier to handle and improves readability of the transports.

Reviewed By: beauby

Differential Revision: D24443496

fbshipit-source-id: 9aa707a6b5a42744c7744c18022d0432aabc37e7