Don't wrap Sockets in shared_ptrs
Summary:
The Socket class is a RAII helper which already has a state that represents the "empty" state, i.e., owning no socket. This is when the fd equals -1 (like nullptr for smart pointer classes) and it's the default-constructed state and also the moved-out state (in fact all movable RAII class must have such a state).

Also, we never need shared ownership of sockets: there's always one class owning them and they are passed around.

So, there's no need to wrap them in shared_ptr (and pay the extra complexity and the extra heap allocation): we can store and pass them by value.

Reviewed By: beauby

Differential Revision: D24336673

fbshipit-source-id: fa5c72af7ae7f77a07ed37d11fdf2a92c5286da0