Move common deferred execution interface to own class
Summary:
This diff is a first step towards "standardizing" the threading model we're using throughout the codebase, by formally defining it as an abstract class. This model consists in the idea that all operations on an object should be "deferred" so they can be run in an event-loop-type way and thus can be reasoned about as if they were single-threaded.

Note that this does not mean we are enforcing this kind of threading model: it's still only "opt-in" and future transports can choose to go a different way if they want to. It's just so that transports that choose to do things this way have a "common language".

Reviewed By: beauby

Differential Revision: D24443497

fbshipit-source-id: 05d30be4b9963c4970ef571cccaf7bb9f1775e1e