Deduplicate SHM and IBV read/write operations
Summary: Due to IBV being a copy of SHM, their ReadOperation and WriteOperation classes were identical. They are some very specialized classes, so the reusability might be limited, but they are completely independent from the rest of the transports (they just depend on the ringbuffer). So I'm moving them to common in order to at least deduplicate them.

Reviewed By: beauby

Differential Revision: D24506737

fbshipit-source-id: 31ac47e8d837a8ecf6a9ffc0f5fcd2fb7059b36e