Introduce cuda.h for common cuda helpers.
Summary:
Extract CUDA helpers, such as the CudaEvent RAII wrapper, helper methods like `cudaDeviceForPointer()`, and logging macro `TP_CUDA_CHECK()` to a separate header so that they can be shared between all CUDA channels.
Moreover, this diff adds a parameter to the constructor of `CudaEvent` to toggle the `cudaEventInterprocess` flag.

Reviewed By: lw

Differential Revision: D24834735

fbshipit-source-id: 1cc3fa88182f045551a925c01af66b906a875e77