Remove obsolete headers and rename some files
Summary:
The connection and listener were now simple aliases and in fact no one used them. So we can remove them and the whole file with them.

This leaves us in a situation where we have only connection_impl.h and connection.cc, with connection.cc containing only methods of ConnectionImpl. That seemed silly, so I renamed connection.cc to connection_impl.cc, and did the same for the listeners. This makes it hopefully easier to guess in which file each class lives, and restores the "convention" of having pair of .h and .cc files with the same name.

To complete this picture, I also split context.cc into context_impl.cc, containing respectively the methods of Context and ContextImpl. In addition to the reasons above, this also helps readability as the methods of Context are just boilerplate (same for all transports, and it's thus best to get them out of the way) whereas the methods of ContextImpl is where the real fun is.

Reviewed By: beauby

Differential Revision: D24950341

fbshipit-source-id: bffcc6aa7b5fed9e64412a3484acf94253290b66