Lock down who can create listener/connection boilerplates
Summary: I was a bit worried that someone writing a new transport which uses the boilerplate might be unaware of the helpers needed to create listeners/connections and attempt to create them directly, thus skipping the bookkeeping code (logging, naming, ...). To prevent that I'm here adding a passkey, which one needs to provide in order to create a connection/listener impl boilerplate, which can only be created by a context impl boilerplate. This should force people to go through those helpers.

Reviewed By: beauby

Differential Revision: D24859088

fbshipit-source-id: c4d4b59312d13e9dcfda27e4354fb03dc0a04e0b