Deduplicate transport boilerplate: connection
Summary:
All our transports have a lot of boilerplate code to deal with common chores such as deferring all incoming operation to the executor loop (so that they are run serially), handling the lifetime of the impl classes and their asynchronous destruction, tagging requests and callbacks with a sequence number and logging their events.

We've been wanting for a long while to deduplicate this code across all our transports, but that was quite hard due to their rather entangled nature (several classes (context, listener, connection), each with a sort-of pimpl pattern, all with mutual interactions). This diff is a first concrete attempt at solving part of that problem, starting with the connections.

It creates an intermediate class in the connection hierarchy (so, it's a subclass of the abstract connection, but a superclass of the concrete per-transport connection). This means that we don't change the abstract API (and thus we don't mandate this approach as the only way of writing a new transport) but we allow any transport to opt-in to it. In fact it's two companion classes: the connection and its impl. Due to the need to store a pointer to the concrete subclass, and the fact that both the intermediate class and the concrete one need to extend enable_shared_from_this, we need to use the curiously recurring template pattern, which is a bit odd.

I'm submitting this diff for consideration and to gather feedback. It still requires some final polishing (choose better names for classes and methods, add comments and docs, avoid storing the pointer to the context impl twice, ...). If we like this approach (or if we settle on another one) I'll extend this approach to all transport (for now I only demonstrate it on UV) and to all classes (context and listener).

Reviewed By: beauby

Differential Revision: D24726530

fbshipit-source-id: b5b2daea4aabbb1c60679014cd0260ec8892811c