Avoid passkey idiom in UV
Summary:
We had a passkey pattern in the connection and the listener to prevent anyone except the context and the listener from creating these classes (while still being compatible with make_shared). However, since the headers that define these classes have now become private, there's an even stronger guarantee that users won't be able to create them. Hence we can drop the passkey and simplify a bit the code.

PS: don't celebrate too early, the passkey is making a comeback, for a different reason, in D24859088

Reviewed By: beauby

Differential Revision: D24859092

fbshipit-source-id: c1e75f941fd49d62cbac4648e845847e0b30fd59