Deduplicate transport boilerplate: context
Summary:
Do to contexts what, in the previous diffs, we did to connections and listeners.

In fact, though, there's a difference, because a transport's context header is public (while its listener and connection header is private) and thus it's harder for us to include private headers in it, including a context_boilerplate.h header. Therefore the context is still manually defined, although all it does is forward calls to its impl.

The impl, on the other hand, benefits from the boilerplate. For now such boilerplate is minimal, because contexts are simple, but this is likely to change soon (e.g., to add the ability to mark an entire context as being in error, to deal with viability, ...). That's when this boilerplate deduplication will prove very useful.

Reviewed By: beauby

Differential Revision: D24859091

fbshipit-source-id: e1452f823931e099c3893e1ca6a9307b41504e13