Add libnop polyfill to boilerplates
Summary:
In order to allow SHM and IBV to provide their own optimized overrides for reading/writing libnop objects, we need to add the necessary "piping" in the boilerplate classes. This piping needs to also provide a polyfill for UV to fall back onto, as it doesn't have any libnop-specific code.

In D24951385, once all transports have migrated, we'll remove the old polyfill, so in effect this stack just moves the polyfill around.

Reviewed By: beauby

Differential Revision: D24951388

fbshipit-source-id: a3bb632405f4128c1a78dfa013e7eeb598cbf27e