In UV, rely on libuv to keep requests alive, instead of "leaking" them
Summary:
Continuing the work from the last diff, we now have a "chain" of ownership that guarantees lifetime: context owns listener/connection, which owns libuv handle. What's left out is the libuv requests. In principle each of them is tied to the lifetime of a libuv handle (e.g., a write request is bound to a libuv stream and, for example, if the stream is closed the request is aborted).

So, one approach could be to add a container to each libuv handle wrapper and store its requests in it. However libuv is basically doing it for us, as it keeps track of which requests belong to a handle and aborts them before closing the handle. This means that the lifetime of the request will always be less than that of the handle, and we can rely on that.

In practice this means that we can "delegate" the ownership of the request to libuv itself, basically "forgetting" about that allocation in our own code (and, in some sense, leaking it, for real). And then, when libuv invokes the callback, we re-acquire ownership, and dispose of the object.

In our code, we can do this nicely and effectively using unique_ptrs. However at the boundary with libuv we need to do two "dirty" steps during the handover: release the unique_ptr, and then re-create it from a raw pointer. This works as intended, but it's a moment of "risk", as if something goes wrong we may lose track of that allocation and leak it for real. Any suggestions on how to make that code more robust are welcome.

Reviewed By: beauby

Differential Revision: D25061001

fbshipit-source-id: 7af03ee137d745b06cd946713c1d718e11502835