In UV, fail when closing handle twice, rather than silently tolerate
Summary: I think we had that "tolerance" from back in the days where we couldn't reliably guarantee that we'd only attempt to close objects once. We now do have such a guarantee (or at least we think we do), so let's make it a loud error if we ever violate such assumption.

Reviewed By: beauby

Differential Revision: D25061011

fbshipit-source-id: 859c48343e634645aa5f0c7856230f8f0a4ddac4