In UV, avoid "leaking" listeners/connectons/handles
Summary:
This is a continuation of what I mentioned in D25061010, in terms of better management of lifetime and ownership.

Now that UV's listeners and connections can reliably manage their lifetime by (un)enrolling with the context, they can in turn provide reliable lifetime to their libuv handles, and thus neither of these objects needs the "leak" monstrosity: the listener/connection will just keep everything alive (themselves and their handles) until the point where we'd unleak, and that has the same effect.

And as a veeery nice plus, by avoiding the need to leak, we avoid the need to refcount these objects, and thus can move from shared_ptrs to unique_ptrs to own libuv handles.

Note that the libuv requests (the other type of libuv resources) doesn't benefit from this, so they will keep using this "leak" hack, for now. (That is, until the next diff in this stack)

Reviewed By: beauby

Differential Revision: D25061003

fbshipit-source-id: ce0b896c63ba99fe201c4706ce141dd2c1ed64b4