Avoid wrapping std::function with optional
Summary: std::function is a RAII class and thus already has an "empty" state (encoded through nullptr). It's therefore redundant to wrap it in an optional, as then there's two empty states: nullopt and optional(nullptr). I guess I wasn't fully aware of this when I first wrote this code...

Reviewed By: patricklabatut

Differential Revision: D25061006

fbshipit-source-id: a1eb56da9c1da7b9fb4638596d5fa67252fec988