Use transport contexts to keep alive connections/listeners
Summary:
This is the first step of a series of changes I want to make to have a more clearly defined lifetime and ownership management in transports. At this stage, this mostly helps the UV transport, where we had a horrific "leak" hack to keep objects alive, where we'd create reference cycles on purpose.

This changes introduces an explicit container of listeners and connections in each context, with each such object enrolling itself at creation and unenrolling itself at destruction (well, rather at error). Hence the context will keep the objects alive even after the user destroyed the last reference to them, allowing the objects to perform an async cleanup if they so need.

I chose to put the enrollment/unenrollment step in each concrete transport, rather than in the boilerplate, because the unenrollment can occur at different times (notably UV will only do it once the close callback fires). And, since each object has a "symmetry" between initImplFromLoop and handleErrorImpl (as if they were constructor and destructor) I chose to also keep the enrollment there.

In the future, I also expect this to provide the backbone on which to implement a replacement for the ClosingEmitter/Receiver (this in turn should improve lifetime of dependent objects, especially in corner cases like objects created after the context gets closed, or at about the same time (racing with it)).

Reviewed By: beauby

Differential Revision: D25061010

fbshipit-source-id: 439798f663cdbb5b221e5544297c95573307f8fb