Enforce naming conventions through clang-tidy
Summary:
I found out that clang-tidy also had rules that could enforce particular naming conventions (casing, prefixes, ...) for certain kinds of names. We were already being quite consistent with such conventions, but I do remember a few occasions in which I had to point out discrepancies during reviews. It would be nice if we could keep such a consistency and if we could have a tool enforce it instead of me.

This is the first batch of rules I'm enabling, which already covers a fair bit of code. I'll run these rules in the next diff and you'll see that not so many changes will be needed. I have left out some rules yet (e.g., one on `using name = type`) because there we have been less consistent. I plan to address that in a later dedicated diff.

Reviewed By: beauby

Differential Revision: D25196336

fbshipit-source-id: be6a649efd9fcf91820a3c5339d8d58b83d09dbf