Run linter to reformat names
Summary:
I'm running the clang-tidy rules I enabled in D25196336 and accepting their suggested fixes. Mostly it's around removing trailing underscores (there were a few that sneaked back in after I thought I had removed them all), changing case, etc. I tweaked a few of the changes manually, for example because there were members that were intended to be private but actually weren't (and this diff highlighted that).

I excluded two files from this run: optional.h (since it's supposed to be a polyfill for std::optional and should follow the stdlib conventions, which are different than ours) and ibv_lib.h (which replicates a lot of libibverbs types, which also follow different conventions). Once we enable clang-tidy to run on the entire codebase at each commit we need to either keep excluding these files (and risk missing real issues in them) or add a lot of `// NOLINT(...)` to them.

Reviewed By: beauby

Differential Revision: D25196335

fbshipit-source-id: c3afe916119619fcdd0229a2af0523a2c6582e1d