Fix PyTorch CI failures due to assert in transports
Summary:
Our API for transports tried to enforce the guarantee that callbacks (of reads/writes on connections, of accepts on listeners) were invoked in the order in which those operation were scheduled. And we had an assert for that in the boilerplate. However, there was at least one case where we failed to honor this: in the UV transport, if we had a bunch of writes queued up, we would abort them if the connection encountered an error (or got closed), but "completing" the abort takes one loop iteration, so we wait to invoke the callbacks until that happens. Yet, if a new write request comes in in the _same_ loop iteration, we call its callback immediately, which could thus be out-of-order, and triggered the assert.

The "proper" fix would be to stash these newer callbacks until all the older ones have fired. However, another "lazier" fix is to give up that ordering guarantee, and remove the assert. This requires no changes to anything else, because the pipe is very robust to out-of-order callbacks (thanks to its state machine), and in turn the pipe has its own guarantees of ordering of callback, thus the API that is presented to end users (i.e., PyTorch) remains completely unchanged.

We could later consider relaxing this ordering guarantee in the pipe too (for other reasons: performance, avoiding head-of-line blocking, ...). But that's a separate discussion.

The failures caused by this assert have showed up in the latest submodule update in PyTorch. I don't see why, as these asserts aren't new. They were affected by the recent introduction of the transport boilerplate, which might have changed things. We'll see, after landing this, if these failures go away or if they were just "hiding" other failures.

Reviewed By: beauby

Differential Revision: D25397944

fbshipit-source-id: dc17338c3d97e9c0c8a4e24be030fc5d2610728f