Remove unused params/methods from ibverbs helpers
Summary: The main point is that transitionIbvQueuePairToReadyToSend was taking a IbvSetupInformation argument which it didn't need, so I'm removing it. I then also removed the field of the IBV connection that stored that IbvSetupInformation. While doing all of that I noticed that a few arguments could be const refs, so I made them so.

Reviewed By: patricklabatut

Differential Revision: D25370398

fbshipit-source-id: 1ae74e32788929bbd2810c21f0c988b025b3fff1