Add query method to CudaEvent RAII wrapper
Summary:
This will be needed for the busy-polling loop of the upcoming CUDA_GDR channel.

Also make it non-copyable/non-movable. (It could in principle be movable, but then it requires us to deal with the "moved-from" empty state. It would be simpler if this used a unique_ptr underneath). I got bitten by this in CUDA_GDR, where I called query on a copy of an event, whose underlying cudaEvent_t had thus already been destroyed.

Reviewed By: patricklabatut

Differential Revision: D25370401

fbshipit-source-id: 36c548efe2601b7df7a714ecc567abe991e752a7