Fix cuda_ipc channel for off set buffers. (#248)
Summary:
In the cuda_ipc channel, we use CUDA's IPC tools to export a buffer's address from one process to the other.
The `cudaIpcGetMemHandle()` function returns a handle to the _base address of the allocation_ of the given pointer, so we need to get the offset of the buffer from the base address and send that over to the other process.
That is only possible using the CUDA driver API, which means we have to link against it as well.

Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/248

Reviewed By: lw

Differential Revision: D25464849

Pulled By: beauby

fbshipit-source-id: 69860b99c1ba20a142205764395fc5af9e5ce004