Make sure we don't release UV resources while work is still coming in
Summary:
### Stack summary

Recent refactorings started causing failures in PyTorch when updating the submodule. Even after finding the culprit commit, the root-cause wasn't the clear. It was somewhat related to changing the place in which various steps of shutdown were done, and thus their relative order. While trying to repro it in a test, even more shutdown-related issues came up, and in this stack I'm trying to solve all the ones I found. The guiding principle is that we should try to make closing a context an "atomic" operation, which instantly sets all child objects (listeners and connections) to error, and then ensures that all operations that follow will honor this error state and won't attempt to further use the context or other resources.

### Commit summary

This might have been the "core" failure that we were seeing in PyTorch. By closing the uv_async_t handle (which we use to wake up the loop when new functions are deferred to it) before we actually handed over control to the on-demand deferred executor meant that new work could come in and try to use that handle. This was a problem. The proper fix is closing it after the hand-over is complete.

Reviewed By: beauby

Differential Revision: D25495681

fbshipit-source-id: 85e5ed0aff67a8e7349e5bd7a28b491b0791ef18