Don't init connections/listeners if context is closed
Summary:
### Stack summary

Recent refactorings started causing failures in PyTorch when updating the submodule. Even after finding the culprit commit, the root-cause wasn't the clear. It was somewhat related to changing the place in which various steps of shutdown were done, and thus their relative order. While trying to repro it in a test, even more shutdown-related issues came up, and in this stack I'm trying to solve all the ones I found. The guiding principle is that we should try to make closing a context an "atomic" operation, which instantly sets all child objects (listeners and connections) to error, and then ensures that all operations that follow will honor this error state and won't attempt to further use the context or other resources.

### Commit summary

In UV, for example, the listeners and connections would create and initialize a libuv handle in their initFromLoop. However, the uv_loop could be closed and gone by then, and this was causing errors. Other transports probably suffered from similar issues. A generic solution is for the boilerplate to detect this and not even invoke initFromLoop. In a sense, we're "shortcutting" initFromLoop+closeFromLoop by not calling either of them and immediately transitioning to error.

Reviewed By: beauby

Differential Revision: D25495682

fbshipit-source-id: 82217ede6c4468179014efe126953579d83e113e