Rely on enrollment to propagate closing from ctx to conn/list
Summary:
And thus retire the ClosingEmitter/Receiver from transports!

### Stack summary

Recent refactorings started causing failures in PyTorch when updating the submodule. Even after finding the culprit commit, the root-cause wasn't the clear. It was somewhat related to changing the place in which various steps of shutdown were done, and thus their relative order. While trying to repro it in a test, even more shutdown-related issues came up, and in this stack I'm trying to solve all the ones I found. The guiding principle is that we should try to make closing a context an "atomic" operation, which instantly sets all child objects (listeners and connections) to error, and then ensures that all operations that follow will honor this error state and won't attempt to further use the context or other resources.

### Commit summary

The ClosingEmitter/Receiver were a weird relic that used weak pointers and their own internal on-demand deferred executor in order for the context to signal to the connections/listeners that they had to close. It was very hard to reason about it. Now, with the boilerplate and the other recent changes, we have each context already expose a neatly-reusable deferred executor, and we have a strong shared_ptr held by the context to each active connection/listener through the enrollment. Leveraging these to signal the closure makes it much easier to see what's going on and ensure that closing connections and listeners happens atomically with the context.

Reviewed By: beauby

Differential Revision: D25495914

fbshipit-source-id: 6a40be79cd675dd1ea546c6bbeaadaa1f083afd4