Make transport listeners init accepted connections atomically
Summary:
### Stack summary

Recent refactorings started causing failures in PyTorch when updating the submodule. Even after finding the culprit commit, the root-cause wasn't the clear. It was somewhat related to changing the place in which various steps of shutdown were done, and thus their relative order. While trying to repro it in a test, even more shutdown-related issues came up, and in this stack I'm trying to solve all the ones I found. The guiding principle is that we should try to make closing a context an "atomic" operation, which instantly sets all child objects (listeners and connections) to error, and then ensures that all operations that follow will honor this error state and won't attempt to further use the context or other resources.

### Commit summary

Connections that are accepted through a listeners are typically half-initialized in the listener (e.g., the socket is opened and accepted there) but then fully-initialized in the the connection's initFromLoop method. Until now, initFromLoop was deferred to later, thus allowing for an error to happen in between, which would mess things up. This commit changes the listener boilerplate so that it calls initFromLoop inline and thus initializes the connection atomically.

Reviewed By: beauby

Differential Revision: D25495685

fbshipit-source-id: f0f9852a7ef362307f0823cbcd12b69455c0bf34