Ensure transport context closing doesn't race with connections/listeners initing
Summary:
### Stack summary

Recent refactorings started causing failures in PyTorch when updating the submodule. Even after finding the culprit commit, the root-cause wasn't the clear. It was somewhat related to changing the place in which various steps of shutdown were done, and thus their relative order. While trying to repro it in a test, even more shutdown-related issues came up, and in this stack I'm trying to solve all the ones I found. The guiding principle is that we should try to make closing a context an "atomic" operation, which instantly sets all child objects (listeners and connections) to error, and then ensures that all operations that follow will honor this error state and won't attempt to further use the context or other resources.

### Commit summary

Even though the closed_ flag is an atomic, it is toggled by the user thread without any synchronization with the event loop. In some tests I've seen a connection first pass the check of the context still being open, but then finding it to be closed immediately after when creating a handle. This was due to this race.

The "fix" is to defer setting the flag to the loop. I suspect that this means the flag at this point could become a *non-atomic* boolean. However, I'll do such a change later (shortly), as I also want to replace that boolean flag with an Error field, which is more generic.

Reviewed By: beauby

Differential Revision: D25495684

fbshipit-source-id: a5160f74276044f5dbb67beea11b35992397c4b3