Lazy loading of libcuda.so. (#252)
Summary:
Linking against `libcuda.so` is problematic for PyTorch, since a CUDA-enabled build of PyTorch is expected to run on machines without GPUs (on which `libcuda.so`is usually not available, while `libcudart.so` gets packaged as a dependency of PyTorch).

This PR attempts to `dlopen(libcuda.so)` from within the `tensorpipe::channel::cuda_ipc::Context::Impl()` constructor, pretty much the same way the ibv transport does. If successful, the required symbols (mainly `cuMemGetAddressRange()`[1]) are exposed on a `CudaLib` object, available through the `Context::Impl::getCudaLib()` method. If `dlopen()` fails, the context is marked as non-viable.

[1] The actual symbol is `cuMemGetAddressRange_v2`, which is what `cuMemGetAddressRange` actually gets `#define`d to in `cuda.h`, while the latter corresponds to an older version, which leads to a `CUDA_ERROR_INVALID_CONTEXT`. This took a while to figure out.

Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/252

Reviewed By: lw

Differential Revision: D25579323

Pulled By: beauby

fbshipit-source-id: 8dbf95f67a439652a7d702de44db31cbce78ab7d