Add sequence numbers to verbose logging in CUDA basic channel
Summary:
When I ported CUDA basic to the boilerplate, clang-tidy made a great point: we weren't using the sequenceNumber argument in the sendImplFromLoop and recvImplFromLoop functions. We definitely should, as otherwise all logging messages will look alike and be less useful.

Note that we also were missing log messages at all for recv on the CPU channel.

Reviewed By: beauby

Differential Revision: D25636928

fbshipit-source-id: 04a18d96ba421b2def20215677a1208567109cdc