Add a GPUDirect channel (InfiniBand for CUDA) (#255)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/255

This is a re-land of D25355282 (https://github.com/pytorch/tensorpipe/commit/d146c1f134f9be3e6ca44214277ea467e254fb00).

The CUDA_GDR channel works as follows:
- Upon creation of the context, we receive from the user a mapping from GPU indices to NIC device names. (In a later diff I'll make this automatic, but to keep this diff simple I'm asking the user to provide it). The context then opens each of the given devices.
- Upon creation of a channel, the two ends first exchange how many NICs they each have, and then open a queue pair between all pairs of NICs (in order to support transfers between any pair of GPUs). This could be done lazily, but for simplicity for now it's eager.
- When sending a tensor, we in fact only store the GPU index in the descriptor, and then wait for the receiver to tell us they are ready to receive.
- On the receiver side, we wait for the target tensor to be ready (by adding an event to the stream and having a callback fire when the event becomes ready). At that point, we register the tensor with libibverbs (so the NIC can write to it directly). Then we send an ibverbs recv request, and notify the sender that we're ready to receive.
- When the sender receives this notification, they go through similar steps: they also wait on the event to become ready and they also register the memory with libibverbs. Then they post an ibverbs send request.
- When these requests finish, each side calls the user callback.

For both the ibverbs polling and the wait on the CUDA events we have a busy polling loop in the context. We could consider replacing this busy polling loop with an implementation that goes to sleep (e.g., like we do in CudaLoop), but we'd need to check that it doesn't impact latency. A simpler change we could do is put the loop to sleep while there's no active requests. (Channels have this luxury, as all their transfers are solicited, whereas transports cannot do it... :( )

Also worth noting that registering CUDA memory with libibverbs can be a veeery slow operation, and doing so in the critical path is a bad idea. My "excuse" is that we cache these registrations, and avoid re-doing them if we already have them. Crossing our fingers that users will typically send the same tensors (or the same storage) over and over, this should amortize the cost pretty well over time.

Reviewed By: beauby

Differential Revision: D25637880

fbshipit-source-id: 8bca80335d5b048ca778b91c45941af94b739885