Fix segfault in CUDA GDR during shutdown
Summary: During `handleError` the channel transitions each `queuePairs_[i][j]` to error, for `i \in [0, numLocalNics_)` and `j \in [0, numRemoteNics_)`. However `numRemoteNics_` is only set during the handshake, and if the channel is closed before that then `numRemoteNics_` will still have an uninitalized value, which if non-zero causes `handleError` to access an invalid element and segfault.

Reviewed By: beauby

Differential Revision: D25828989

fbshipit-source-id: 7585a87ddc7b41964fc10f44f4c9377532588c45