Avoid uncaught exception causing "fork bomb" in ProcessPeerGroup
Summary: For CUDA tests we need to run each "node" of a test as a separate child subprocess, hence we fork. At the end of the branch for the child process we exit. However, if the child code raises an exception, this exit is skipped, but then gtest would catch the exception, and at that point it would continue executing as if it was the parent. Hence it would go on and itself launch further subprocesses, which may also throw, and then they would also go on and think they are the parents, and in turn fork, etc.

Reviewed By: beauby

Differential Revision: D25825121

fbshipit-source-id: fc062a2554338f4aa08fb9f3cc205fdad19196fa