Always create channel ctxs in subprocesses to avoid initializing CUDA
Summary:
CUDA GDR will soon become the first channel that performs CUDA calls inside its Context constructor. This means that creating a Context will initialize CUDA and thus "pollute" the process, as any later child of that process will be unable to re-initialize CUDA.

The DomainDescriptor test was the only test of the CUDA suite that did not spawn a subprocess when it created a Context. Thus any other test case running after that one would fail.

The fix here is sloppy, as instead of implementing a "proper" solution to this I'm reusing the PeerGroup helpers that we already had, but which were designed for another use case. They work fine, but perhaps we'll want to revisit this and invest in something nicer.

Reviewed By: beauby

Differential Revision: D25846287

fbshipit-source-id: ca5be0e7786004736a4ea339cccd78ef222206c8