Reuse pinned buffers in cuda_basic.
This PR introduces a `CudaPinnedMemoryAllocator` (shared among
`cuda_basic` channels from a same context) that pre-allocates one
large slab of host pinned memory, divided into chunks that can be
used by channels when they need to stage data to host buffers.

The proposed strategy is to return a whole existing chunk when one is
available and the requested buffer is at most the chunk
size. Otherwise, we fall back to allocating a new temporary pinned
buffer.
We could consider copying a large device tensor to multiple
non-contiguous chunks but it is not clear there would be any perf
benefits of exchanging one `cudaMallocHost()` call for multiple
`cudaMemcpyAsync()`.
One optimization that would make sense would be to use several
contiguous chunks when possible.