Update clang-format version used by CircleCI (#261)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/261

Facebook's internal version of clang-format got updated to v11 over the weekend and this apparently changed the formatting. I think this was a "bugfix", because the `AlignAfterOpenBracket`, which we've set to `AlwaysBreak`, was supposed to also cover curly brackets, but it seems it didn't, whereas now it does. I "bisected" this and it appears the change was introduces in clang-format v10. This is good because v11 isn't available on Ubuntu 18.04, which is what we use in CircleCI, whereas v10 is.

Reviewed By: patricklabatut

Differential Revision: D25867695

fbshipit-source-id: ae98b9e9befee3c86a28820988b9fc3735557e52